Updated to use react
