--- smart-scheduler/README.md
@@ -1,11 +1,0 @@
-## Welcome ACM App Development Club Members
-
-## Instructions to run the project
-- Type 'npm i' in your console
-- Type 'npm start' in your console
-- Choose where you will simulate the app and choose that option in the console
-
-## Instructions for collaboration
-- Choose an issue on gitlab
-- Get to crackin
-- If there is any ambiguity, ask any of the other collaborators
