--- smart-scheduler/app/index.tsx
@@ -1,6 +1,0 @@
-// Serves as the entrypoint to the app
-import { Redirect } from 'expo-router';
-
-export default function Index() {
-  return <Redirect href="/home" />;
-}
