--- smart-scheduler/constants/Colors.ts
@@ -1,19 +1,0 @@
-/**
- * Below are the colors that are used in the app.
- * This setup aligns with the app's agreed-upon color scheme for dark mode only.
- */
-
-export const Colors = {
-  dark: {
-    background: '#1C1C1E',       // Main background color, very dark grey, almost black
-    text: '#E5E5E7',              // Primary text color, light grey-white for readability
-    secondaryText: '#A3A3A5',     // Secondary or muted text, medium grey for subtle contrast
-    tint: '#4A90E2',              // Accent color for buttons, highlights, and interactive elements, cool blue
-    icon: '#4A90E2',              // Icons matching the accent color
-    tabIconDefault: '#A3A3A5',    // Medium grey for default icons
-    tabIconSelected: '#4A90E2',   // Accent color for selected icons
-    error: '#FF453A',             // Vibrant red for error messages, warnings, and critical alerts
-  },
-};
-
-export default Colors;
